Move "everything" into core
